WIP @emilyagras #11 Mid webpack install, I've gota goto sleep
