+++ frontend/package.json
@@ -8,18 +8,13 @@
     "babel-cli": "^6.24.1",
     "babel-jest": "^20.0.3",
     "babel-preset-env": "^1.5.1",
-    "eslint-config-airbnb": "^14.1.0",
-    "eslint-config-airbnb-base": "^11.1.3",
-    "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-jsx-a11y": "^4.0.0",
-    "eslint-plugin-react": "^6.10.3",
     "jest": "^20.0.4",
     "nodemon": "^1.11.0",
-    "tap": "^10.3.2",
     "ws": "^3.0.0"
   },
   "scripts": {
-    "test": "jest"
+    "test": "jest",
+    "build" : "./node_modules/.bin/babel src -d dist"
   },
   "repository": {
     "type": "git",
